requirements.txt
streamlit
plotly
pandas
PyPDF2
python-dotenv
dashscope
streamlit-authenticator
bcrypt
pyyaml
pinecone